@@ -34,22 +34,22 @@ export class UsersManagementComponent implements OnInit, AfterViewInit {
     allRoles: Role[] = [];
 
 
-    @ViewChild('indexTemplate')
+    @ViewChild('indexTemplate', { static: true })
     indexTemplate: TemplateRef<any>;
 
-    @ViewChild('userNameTemplate')
+    @ViewChild('userNameTemplate', { static: true })
     userNameTemplate: TemplateRef<any>;
 
-    @ViewChild('rolesTemplate')
+    @ViewChild('rolesTemplate', { static: true })
     rolesTemplate: TemplateRef<any>;
 
-    @ViewChild('actionsTemplate')
+    @ViewChild('actionsTemplate', { static: true })
     actionsTemplate: TemplateRef<any>;
 
-    @ViewChild('editorModal')
+    @ViewChild('editorModal', { static: true })
     editorModal: ModalDirective;
 
-    @ViewChild('userEditor')
+    @ViewChild('userEditor', { static: true })
     userEditor: UserInfoComponent;
 
     constructor(private alertService: AlertService, private translationService: AppTranslationService, private accountService: AccountService) {
